requirements.txt
amqp==5.1.1
asgiref==3.6.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.5
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.14.0
django-timezone-field==5.0
djangorestframework==3.14.0
flower==1.2.0
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
humanize==4.6.0
idna==3.4
kombu==5.2.4
lxml==4.9.2
oauthlib==3.2.2
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stripe==5.4.0
tornado==6.3.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
